Move image Load/Write shards
