+++ shards/modules/imaging/imaging.cpp
@@ -1,16 +1,27 @@
 /* SPDX-License-Identifier: BSD-3-Clause */
 /* Copyright © 2020 Fragcolor Pte. Ltd. */
 
+#include "imaging.hpp"
+#include <shards/modules/core/file_base.hpp>
 #include <shards/core/module.hpp>
 #include <shards/core/shared.hpp>
 #include <shards/core/params.hpp>
+#include <fstream>
 
 #include "linalg.h"
 using namespace linalg::aliases;
 
 #define STB_IMAGE_RESIZE_IMPLEMENTATION
+#define STB_IMAGE_IMPLEMENTATION
+#define STB_IMAGE_WRITE_IMPLEMENTATION
 #include <stb_image_resize.h>
-#include "imaging.hpp"
+#include <stb_image.h>
+#include <stb_image_write.h>
+
+// Thanks windows for polluting the global namespace
+#ifdef LoadImage
+#undef LoadImage
+#endif
 
 namespace shards {
 namespace Imaging {
@@ -490,16 +501,250 @@ struct ImageGetPixel {
 
     if (pixsize == 1) {
       return (bool)*_asInteger ? convert<uint8_t>(image, coord, Conv_CastInt4) //
-                              : convert<uint8_t>(image, coord, Conv_UIntToFloat4);
+                               : convert<uint8_t>(image, coord, Conv_UIntToFloat4);
     } else if (pixsize == 2) {
       return (bool)*_asInteger ? convert<uint16_t>(image, coord, Conv_CastInt4) //
-                              : convert<uint16_t>(image, coord, Conv_UIntToFloat4);
+                               : convert<uint16_t>(image, coord, Conv_UIntToFloat4);
     } else if (pixsize == 4) {
       return (bool)*_asInteger ? convert<float>(image, coord, Conv_FloatToInt4) //
-                              : convert<float>(image, coord, Conv_CastFloat4);
+                               : convert<float>(image, coord, Conv_CastFloat4);
     } else {
       throw std::logic_error("Invalid image format");
     }
+  }
+};
+
+struct LoadImage : public FileBase {
+  enum class BPP { u8, u16, f32 };
+  DECL_ENUM_INFO(BPP, BPP, 'ibpp');
+
+  static SHTypesInfo inputTypes() { return CoreInfo::BytesOrAny; }
+  static SHTypesInfo outputTypes() { return CoreInfo::ImageType; }
+
+  static inline Parameters params{FileBase::params,
+                                  {{"BPP", SHCCSTR("bits per pixel (HDR images loading and such!)"), {BPPEnumInfo::Type}},
+                                   {"PremultiplyAlpha",
+                                    SHCCSTR("Toggle premultiplication of alpha channels (E.g. To support PNG images)"),
+                                    {CoreInfo::BoolType}}}};
+
+  static SHParametersInfo parameters() { return params; }
+
+  void setParam(int index, const SHVar &value) {
+    switch (index) {
+    case 1:
+      _bpp = BPP(value.payload.enumValue);
+      break;
+    case 2:
+      _premultiplyAlpha = value.payload.boolValue;
+      break;
+    default:
+      FileBase::setParam(index, value);
+    }
+  }
+
+  SHVar getParam(int index) {
+    switch (index) {
+    case 1:
+      return Var::Enum(_bpp, CoreCC, BPPEnumInfo::TypeId);
+    case 2:
+      return Var(_premultiplyAlpha);
+    default:
+      return FileBase::getParam(index);
+    }
+  }
+
+  SHVar _output{};
+  BPP _bpp{BPP::u8};
+  bool _premultiplyAlpha{};
+
+  void cleanup() {
+    if (_output.valueType == SHType::Image && _output.payload.imageValue.data) {
+      stbi_image_free(_output.payload.imageValue.data);
+      _output = Var::Empty;
+    }
+
+    FileBase::cleanup();
+  }
+
+  SHVar activate(SHContext *context, const SHVar &input) {
+    bool bytesInput = input.valueType == SHType::Bytes;
+
+    // free the old image if we have one
+    if (_output.valueType == SHType::Image && _output.payload.imageValue.data) {
+      stbi_image_free(_output.payload.imageValue.data);
+      _output = Var::Empty;
+    }
+
+    stbi_set_flip_vertically_on_load_thread(0);
+
+    uint8_t *bytesValue;
+    uint32_t bytesSize;
+
+    // if we have a file input, load them into bytes form
+    std::string filename;
+    std::vector<uint8_t> buffer;
+
+    if (!bytesInput) {
+      // need a proper filename in this case
+      if (!getFilename(context, filename)) {
+        throw ActivationError(fmt::format("File not found: {}!", filename));
+      }
+
+      // load the file
+      std::ifstream file(filename, std::ios::binary);
+      if (!file) {
+        throw ActivationError(fmt::format("Error opening file: {}!", filename));
+      }
+
+      // get length of file
+      file.seekg(0, std::ios::end);
+      std::streampos length = file.tellg();
+      file.seekg(0, std::ios::beg);
+
+      // read file into buffer
+      buffer.resize(length);
+      file.read(reinterpret_cast<char *>(buffer.data()), length);
+
+      bytesValue = buffer.data();
+      bytesSize = static_cast<uint32_t>(length);
+    } else {
+      // image already given in bytes form
+      bytesValue = input.payload.bytesValue;
+      bytesSize = input.payload.bytesSize;
+    }
+
+    _output.valueType = SHType::Image;
+    int x, y, n;
+    switch (_bpp) {
+    case BPP::u8:
+      _output.payload.imageValue.data =
+          reinterpret_cast<uint8_t *>(stbi_load_from_memory(bytesValue, static_cast<int>(bytesSize), &x, &y, &n, 0));
+
+      _output.payload.imageValue.flags = 0;
+      break;
+    case BPP::u16:
+      _output.payload.imageValue.data =
+          reinterpret_cast<uint8_t *>(stbi_load_16_from_memory(bytesValue, static_cast<int>(bytesSize), &x, &y, &n, 0));
+
+      _output.payload.imageValue.flags = SHIMAGE_FLAGS_16BITS_INT;
+      break;
+    default:
+      _output.payload.imageValue.data =
+          reinterpret_cast<uint8_t *>(stbi_loadf_from_memory(bytesValue, static_cast<int>(bytesSize), &x, &y, &n, 0));
+
+      _output.payload.imageValue.flags = SHIMAGE_FLAGS_32BITS_FLOAT;
+      break;
+    }
+
+    if (!_output.payload.imageValue.data) {
+      throw ActivationError("Failed to load image file");
+    }
+
+    _output.payload.imageValue.width = uint16_t(x);
+    _output.payload.imageValue.height = uint16_t(y);
+    _output.payload.imageValue.channels = uint16_t(n);
+
+    // Premultiply the alpha channel if premultiply option is chosen
+    auto pixsize = getPixelSize(_output);
+    if (_premultiplyAlpha) {
+      // premultiply the alpha channel
+      switch (pixsize) {
+      case 1:
+        Imaging::premultiplyAlpha<uint8_t>(_output, _output, _output.payload.imageValue.width, _output.payload.imageValue.height);
+        break;
+      case 2:
+        Imaging::premultiplyAlpha<uint16_t>(_output, _output, _output.payload.imageValue.width,
+                                            _output.payload.imageValue.height);
+        break;
+      case 4:
+        Imaging::premultiplyAlpha<float>(_output, _output, _output.payload.imageValue.width, _output.payload.imageValue.height);
+        break;
+      }
+    }
+
+    _output.version = 0;
+    return _output;
+  }
+};
+
+struct WritePNG : public FileBase {
+  static inline Types OutputTypes = {{CoreInfo::BytesType, CoreInfo::ImageType}};
+  std::vector<uint8_t> _scratch;
+  std::vector<uint8_t> _output;
+
+  static SHTypesInfo inputTypes() { return CoreInfo::ImageType; }
+  SHTypesInfo outputTypes() { return OutputTypes; }
+
+  static void write_func(void *context, void *data, int size) {
+    auto self = reinterpret_cast<WritePNG *>(context);
+    self->_output.resize(size);
+    memcpy(self->_output.data(), data, size);
+  }
+
+  SHTypeInfo compose(const SHInstanceData &data) {
+    // If param is none we output the bytes directly
+    if (_filename->valueType == SHType::None) {
+      return CoreInfo::BytesType;
+    } else {
+      return CoreInfo::ImageType;
+    }
+  }
+
+  SHVar activate(SHContext *context, const SHVar &input) {
+    auto pixsize = getPixelSize(input);
+
+    std::string filename;
+    if (_filename->valueType != SHType::None) {
+      if (!getFilename(context, filename, false)) {
+        throw ActivationError("Path does not exist!");
+      }
+    }
+
+    int w = int(input.payload.imageValue.width);
+    int h = int(input.payload.imageValue.height);
+    int c = int(input.payload.imageValue.channels);
+
+    // demultiply alpha if needed, limited to 4 channels
+    if (c == 4 && (input.payload.imageValue.flags & SHIMAGE_FLAGS_PREMULTIPLIED_ALPHA) == SHIMAGE_FLAGS_PREMULTIPLIED_ALPHA) {
+      _scratch.resize(w * h * 4 * pixsize);
+      for (int ih = 0; ih < h; ih++) {
+        for (int iw = 0; iw < w; iw++) {
+          switch (pixsize) {
+          case 1:
+            Imaging::demultiplyAlpha<uint8_t>(input, _scratch, w, h);
+            break;
+          case 2:
+            Imaging::demultiplyAlpha<uint16_t>(input, _scratch, w, h);
+            break;
+          case 4:
+            Imaging::demultiplyAlpha<float>(input, _scratch, w, h);
+            break;
+          }
+        }
+      }
+
+      // all done, write the file or buffer
+      if (!filename.empty()) {
+        if (0 == stbi_write_png(filename.c_str(), w, h, c, _scratch.data(), w * c))
+          throw ActivationError("Failed to write PNG file.");
+      } else {
+        if (0 == stbi_write_png_to_func(write_func, this, w, h, c, _scratch.data(), w * c))
+          throw ActivationError("Failed to write PNG file.");
+        return Var(_output.data(), _output.size());
+      }
+    } else {
+      // just write the file or buffer in this case straight
+      if (!filename.empty()) {
+        if (0 == stbi_write_png(filename.c_str(), w, h, c, input.payload.imageValue.data, w * c))
+          throw ActivationError("Failed to write PNG file.");
+      } else {
+        if (0 == stbi_write_png_to_func(write_func, this, w, h, c, input.payload.imageValue.data, w * c))
+          throw ActivationError("Failed to write PNG file.");
+        return Var(_output.data(), _output.size());
+      }
+    }
+
+    return input;
   }
 };
 
@@ -509,6 +754,8 @@ struct ImageGetPixel {
 SHARDS_REGISTER_FN(imaging) {
   using namespace shards::Imaging;
 
+  REGISTER_ENUM(LoadImage::BPPEnumInfo);
+
   REGISTER_SHARD("GetImagePixel", ImageGetPixel);
   REGISTER_SHARD("Convolve", Convolve);
   REGISTER_SHARD("StripAlpha", StripAlpha);
@@ -516,4 +763,6 @@ SHARDS_REGISTER_FN(imaging) {
   REGISTER_SHARD("DemultiplyAlpha", DemultiplyAlpha);
   REGISTER_SHARD("FillAlpha", FillAlpha);
   REGISTER_SHARD("ResizeImage", Resize);
+  REGISTER_SHARD("LoadImage", LoadImage);
+  REGISTER_SHARD("WritePNG", WritePNG);
 }
